Feature: Added hashflix.
This is the last practice, I might delete this
repo since it turned out to be shitty code lol,
I want to redo this in a kinda official basic
library for data structures. I started with a
repo named ckastal which was a copy from this
project but the code is kinda useless lol.

This repo is trash lmao. See you in space cowboy!